Make scrollable list always draw new on setState
That may be bad regarding performance stuff. But with setState we will change the number of items shown in the list. If we adjust the scroll position via a ScrollController before the list is redrawn we might get a wrong position out of it.
By always redrawing the list we don't have to think about that any longer. We can always use the initialScrollPosition for each redraw.